Initial implementation of UI updates for Task Analysis.
